Still trying to get svnwiki stood up...
Found an older version of the list of required Chicken Scheme Eggs that
came closer to working. The big missing dependency is simple-logging.
Current working theory: an older version of Alejandro's Eggs are needed
that don't have the simple-logging dependency. We could possibly get the
older versions, but progress has stopped on the svnwiki project. I think
it's time to look elsewhere for a SCM-backed wiki.